difficultmenu working

+++ avaliacoes/semana007/ex003.c
@@ -11,7 +11,7 @@ void display_table(int matriz[M][N], int interface [M][N]);
 void clearscreen();
 int main_menu();
 void flush_in();
-int difficult_menu();
+float difficult_menu();
 int gScore;
 
 
@@ -42,6 +42,7 @@ int main(){
             players ++;
             gScore = 0;
             endOfGame = 0;
+            
             num_mines = M * N * mines_percentage;
 
             for(i = 0; i < M; i ++){
@@ -219,7 +220,7 @@ int main_menu() {
     return option;
 }
 
-int difficult_menu(){
+float difficult_menu(){
     int option; 
     float mines_percentage = 0.1;
     clearscreen();
